added logger,swagger,exception handling

+++ src/main/java/com/sonu/urlStorage/model/URLInfoResponse.java
@@ -2,10 +2,9 @@ package com.sonu.urlStorage.model;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
 @Data
 public class URLInfoResponse {
     private String url;
     private String shortKey;
-    private BigDecimal count;
+    private long count;
 }
